python_backend/requirements-enhanced.txt
# Core dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
pydantic==2.10.3

# Async database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
redis==5.2.0

# AI/ML
ultralytics==8.3.40
opencv-python==4.10.0.84
numpy==1.26.4
pillow==11.0.0
tensorflow==2.15.0

# Async processing
celery==5.4.0
redis==5.2.0

# Rate limiting & caching
slowapi==0.1.9
fastapi-limiter==0.1.5

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8

# Development & testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-html==4.1.1
pytest-cov==4.1.0
python-dotenv==1.0.1

# Security scanning
bandit==1.7.5
safety==3.0.1
semgrep==1.45.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# File processing
python-magic==0.4.27
aiofiles==23.2.1
